COASTLINES - first upload of new shape files   MAGP-435
All shp files are new, with exception of 10m/10m_land*.shp which are the old one for comparison. First impression is good with "only" the great lakes being a problem. The spurious lines in the coast lines are gone.
added coast line files which we should really use when plotting the coast lines.